Integer to Long

+++ academy-backend/src/main/java/uz/jurayev/academy/controller/FamilyConditionController.java
@@ -29,18 +29,18 @@ public class FamilyConditionController {
     }
 
     @GetMapping("{id}")
-    public HttpEntity<?> getOne(@PathVariable Integer id) {
+    public HttpEntity<?> getOne(@PathVariable Long id) {
         FamilyCondition familyCondition = familyConditionService.getOne(id);
         return ResponseEntity.status(familyCondition != null ? 200 : 409).body(familyCondition);
     }
 
     @PutMapping("/{id}")
-    public HttpEntity<?> edit(@PathVariable Integer id, @RequestBody FamilyCondition condition) {
+    public HttpEntity<?> edit(@PathVariable Long id, @RequestBody FamilyCondition condition) {
         FamilyCondition familyCondition = familyConditionService.edit(id, condition);
         return ResponseEntity.status(familyCondition != null ? 201 : 409).body(familyCondition);
     }
     @DeleteMapping("{id}")
-    public HttpEntity<?> delete(@PathVariable Integer id){
+    public HttpEntity<?> delete(@PathVariable Long id){
         boolean delete = familyConditionService.delete(id);
         if (delete)
             return ResponseEntity.noContent().build();
