+++ academy-backend/src/main/java/uz/jurayev/academy/repository/RegionRepository.java
@@ -8,7 +8,7 @@ import java.util.Optional;
 public interface RegionRepository extends JpaRepository<Region, Long> {
     boolean existsByRegionName(String name);
 
-    Optional<Region> findById(Integer id);
+    Optional<Region> findById(Long id);
 
-    void deleteById(Integer id);
+    void deleteById(Long id);
 }
